requirements.txt
Django==1.11.21
pytz==2019.1
djangorestframework==3.9.0
django-rest-swagger==2.1.2
pytest
pytest-cov
pytest-django
pytest-factoryboy
pytest-mock
pytest-sugar
requests-mock